Final Push
Added backtracking capability in terms of an auto-solver. Needs to be polished with code refactoring, but it works as intended!

You can press "S" to start the auto-solver. If the auto-solver fails, it will alert the user in the log while reverting the game board.
You can delete an entry (one in green) by using backspace, or by simply overriding it.
You can press Ctrl+Z to undo
You can press Ctrl+Y to redo (note: an edit after an undo will override the history, like in google chrome)
You can press Left, Right, Up and Down to move around the grid.
Type in a number 1 - 9 to fill that space (note 0 will skip to the next cell).

This project requires pygame and python 3.7+.